+++ alex_popov/final-project/src/css/main/_about.scss
@@ -2,4 +2,5 @@
     position: relative;
     width: 100%;    
     top: 0;
+    border-bottom: 1px solid rgb(88, 88, 88)
 }
