+++ alex_popov/final-project/src/js/components/Header/Navigation.jsx
@@ -1,14 +1,16 @@
 import React from 'react';
 import { NavLink } from 'react-router-dom';
 import LoginIcon from './LoginIcon';
+import BurgerMenu from './BurgerMenu';
 
-const Navigation = () => {
-    return (
-      <nav className="header-links">
-            <div id='menuIcon' className='header_mobileMenu'>
+const Navigation = () => (
+    <nav className="header-links">
+        <BurgerMenu />
+            {/* <div id='menuIcon' className='header_mobileMenu'>
                 <svg id='menuLogo' xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill="white" width="30px" height="30px"><path d="M0 0h24v24H0z" fill="none"/><path d="M3 18h18v-2H3v2zm0-5h18v-2H3v2zm0-7v2h18V6H3z"/></svg>
-            </div>
-            <div id='header_mobileMenu' className='header_mainNav'>
+            </div> */}
+        <div id='header_mobileMenu' className='header_mainNav'>
+
                 <NavLink 
                     activeClassName='header_link-active' 
                     className='header_link' 
@@ -19,6 +21,7 @@ const Navigation = () => {
                 >
                     Main
                 </NavLink>
+
                 <NavLink 
                     activeClassName='header_link-active' 
                     className='header_link' 
@@ -29,6 +32,7 @@ const Navigation = () => {
                 >
                     services
                 </NavLink>
+
                 <NavLink 
                     activeClassName='header_link-active' 
                     className='header_link' 
@@ -36,6 +40,7 @@ const Navigation = () => {
                 >
                     Booking
                 </NavLink>
+
                 <NavLink 
                     activeClassName='header_link-active' 
                     className='header_link' 
@@ -43,13 +48,13 @@ const Navigation = () => {
                 >
                     contacts
                 </NavLink>
-                {/* <button id='activate'>activate</button> */}
-            </div>
-            <NavLink className='header-link' to='/user' >
+
+        </div>
+
+        <NavLink className='header-link' to='/user' >
                 <LoginIcon />
-            </NavLink>
-      </nav>
-    )
-}
+        </NavLink>
+    </nav>
+)
 
 export default Navigation;
