+++ alex_popov/final-project/src/js/components/Logo.jsx
