+++ alex_popov/final-project/src/js/components/main/Contacts.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Footer from '../Footer/Footer';
+import Footer from '../Footer/Footer.jsx';
 
 function Contacts() {
     return (
